+++ crates/audio/src/lib.rs
@@ -1,6 +1,7 @@
 mod audio;
 mod kira_static_data;
 mod manager;
+mod utils;
 
 pub use audio::Audio;
 pub use manager::AudioManager;